1. This should fix Issue 1625 (Bizarre Auto Frame Limit)    Now the frame limiter yields on CPU thread, not as before on GPU thread mistakenly
2. Fixed clear of VI interrupts
   I guess VI interrupts are not used at all, because they were never cleared before

3. Made GPU thread 0% processor usage when paused whatever your active config is.
   I tried the event approach but somehow the thread resume latency is excessively long (Who can tell me why?)

git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@4790 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Core/Core/Src/OnFrame.cpp
@@ -32,7 +32,7 @@ u32 g_autoFirstKey = 0, g_autoSecondKey = 0;
 bool g_bFirstKey = true;
 PlayMode g_playMode = MODE_NONE;
 
-int g_framesToSkip = 0, g_frameSkipCounter = 0;
+unsigned int g_framesToSkip = 0, g_frameSkipCounter = 0;
 
 int g_numPads = 0;
 ControllerState *g_padStates;
@@ -70,14 +70,13 @@ void FrameUpdate() {
 			EndPlayInput();
 	}
 
-
 	g_bPolled = false;
 }
 
 void SetFrameSkipping(unsigned int framesToSkip) {
 	cs_frameSkip.Enter();
 
-	g_framesToSkip = (int)framesToSkip;
+	g_framesToSkip = framesToSkip;
 	g_frameSkipCounter = 0;
 
 	// Don't forget to re-enable rendering in case it wasn't...
